+++ toolium/test/pageobjects/test_mobile_page_object.py
@@ -37,7 +37,7 @@ def driver_wrapper():
     config_files.set_config_directory(os.path.join(root_path, 'conf'))
     config_files.set_config_properties_filenames('properties.cfg')
     driver_wrapper = DriverWrappersPool.get_default_wrapper()
-    driver_wrapper.configure(tc_config_files=config_files)
+    driver_wrapper.configure(config_files)
 
     return driver_wrapper