+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ v1.4.0
 - Rename config property 'restart_driver_fail' in [Driver] section to 'restart_driver_after_failure'
 - Add @no_driver feature or scenario tag to do not start the driver in these tests
 - Fix output folder names when driver type is empty
+- Fix output log name when `Config_environment` is used
 - Fix Chrome options using remote drivers with Selenium >= 3.6.0
 
 v1.3.0